fix: move suggestions to result object
