+++ src/main/kotlin/no/digdir/fdk/searchservice/model/SuggestionsResult.kt
@@ -1,0 +1,5 @@
+package no.digdir.fdk.searchservice.model
+
+data class SuggestionsResult(
+    val suggestions: List<Suggestion>,
+)
