+++ src/main/kotlin/no/digdir/fdk/searchservice/service/SuggestionService.kt
@@ -4,6 +4,7 @@ import co.elastic.clients.elasticsearch._types.FieldValue
 import no.digdir.fdk.searchservice.model.SearchObject
 import no.digdir.fdk.searchservice.model.SearchType
 import no.digdir.fdk.searchservice.model.Suggestion
+import no.digdir.fdk.searchservice.model.SuggestionsResult
 import org.springframework.data.elasticsearch.client.elc.NativeQuery
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations
 import org.springframework.data.elasticsearch.core.SearchHits
@@ -18,11 +19,11 @@ class SuggestionService(
     private fun suggestResource(query: String, searchType: List<SearchType>?): SearchHits<SearchObject> =
         elasticsearchOperations.search(suggestionQuery(query, searchType), SearchObject::class.java)
 
-    fun suggestResources(query: String, searchType: List<SearchType>?): List<Suggestion> =
-        suggestResource(query, searchType)
-            .map { it.content }
-            .map { it.toSuggestion() }
-            .toList()
+    fun suggestResources(query: String, searchType: List<SearchType>?): SuggestionsResult =
+            SuggestionsResult(suggestResource(query, searchType)
+                .map { it.content }
+                .map { it.toSuggestion() }
+                .toList())
 
     private fun SearchObject.toSuggestion(): Suggestion =
         Suggestion(
